api/requirements.txt
# Dependencias para la Serverless Function de Vercel
openai>=1.50.0
pydantic>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0

